axns

Struct AxNamespace

pub struct AxNamespace { /* private fields */ }

A namespace that contains all user-defined resources.

There are two types of namespaces:

Global namespace: this namespace is globally unique and all threads share the resources in it. Resources are statically collected into the axns_resource section, and the global namespace is constructed by the base address of the section (AxNamespace::global).

Thread-local namespace: this namespace is per-thread, each thread should call AxNamespace::new_thread_local() to allocate a memory area as its namespace. Layout of resources in global and thread-local namespaces is consistent. Each namespace has its own resources, which may be unique or shared between threads by the Arc wrapper.

Implementations

impl AxNamespace

pub const fn base(&self) -> *mut u8

Returns the base address of the namespace, which points to the start of all resources.

pub fn size(&self) -> usize

Returns the size of the namespace (size of all resources).

pub fn global() -> Self

Returns the global namespace.

pub fn new_thread_local() -> Self

Constructs a new thread-local namespace.

Each thread can have its own namespace instead of the global one, to isolate resources between threads.

This function allocates a memory area to store the thread-local resources, and copies from the global namespace as the initial value.