axns

Trait AxNamespaceIf

pub trait AxNamespaceIf {
    // Required method
    fn current_namespace_base() -> *mut u8;
}

The interfaces need to be implemented when enable thread-local namespaces.

Required Methods

fn current_namespace_base() -> *mut u8

Returns the pointer to the current namespace.

It usually needs to be obtained from the thread local storage.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.